+++ packages/analyzer_buffer/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased patch
+
+- Fix `DartType.toCode` with nullable types
+
 ## 0.1.7 - 2025-08-11
 
 - Fixed `DartType.toCode` with typedefs, records, and many other types.